Move command_ard_getter to helper dir
